+++ R/print.env_summarized.R
@@ -25,9 +25,6 @@
 #' x <- summarize_envelope(cluster_env)
 #' print(x)
 #'
-#' @aliases print.env_summarized
-#' @rdname print.env_summarized
-#'
 #' @export
 print.env_summarized <- function(x, return_area = FALSE, digits = 2, ...) {