+++ R/simulate_antecedent_conditions.R
@@ -44,9 +44,6 @@
 #' Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in
 #' ecology. Chapman and Hall/CRC Press, Boca Raton, USA. <isbn:978-1-4200-8254-8>
 #'
-#' @aliases simulate_antecedent_conditions
-#' @rdname simulate_antecedent_conditions
-#'
 #' @export
 simulate_antecedent_conditions <- function(x, i, j, nsim, heterogenous = FALSE, ...) {